[유민형] Player랑 몬스터 총알 소환안되는 버그 및 서로 데미지 부여 기능 구현

+++ 666GunsReborn/Assets/Main/Scripts/Character/Enemy/EnemyStat.cs
@@ -58,7 +58,7 @@ namespace Character.Enemy
         /// </summary>
         /// <param name="damage"></param>
         /// <returns></returns>
-        public bool TakeDamage(int damage)
+        public bool TakeDamage(float damage)
         {
             // 데미지 부여
             _currentHealth -= damage;
